Check proc-macro attributes in ast validation.
